Don't allow account deletion when linked to expenses or incomes

+++ Source/Server/Pot.App/Features/Accounts/Delete/DeleteAccountService.cs
@@ -1,7 +1,9 @@
 ﻿using AllOverIt.Assertion;
 using AllOverIt.Logging.Extensions;
+using AllOverIt.Patterns.Result;
 using Microsoft.Extensions.Logging;
 using Pot.App.Features.Accounts.Create;
+using Pot.App.Features.Accounts.Delete.EntityChecks;
 using Pot.Data.Repositories.Accounts;
 
 namespace Pot.App.Features.Accounts.Delete;
@@ -9,33 +11,38 @@ namespace Pot.App.Features.Accounts.Delete;
 internal sealed class DeleteAccountService : IDeleteAccountService
 {
     private readonly IPersistableAccountRepository _accountRepository;
+    private readonly IPreCreateChecker _preCreateChecker;
     private readonly ILogger _logger;
 
-    public DeleteAccountService(IPersistableAccountRepository accountRepository, ILogger<CreateAccountService> logger)
+    public DeleteAccountService(IPersistableAccountRepository accountRepository, IPreCreateChecker preCreateChecker, ILogger<CreateAccountService> logger)
     {
         _accountRepository = accountRepository.WhenNotNull();
+        _preCreateChecker = preCreateChecker.WhenNotNull();
         _logger = logger.WhenNotNull();
     }
 
-    public async Task<bool> DeleteAccountAsync(Guid id, CancellationToken cancellationToken)
+    public async Task<EnrichedResult<bool>> DeleteAccountAsync(Guid accountId, CancellationToken cancellationToken)
     {
         _logger.LogCall(this);
 
-        // TODO: Don't allow deletion of accounts with associated expenses / incomes.
-        //       Consider soft deletes. ? Reporting requirements.
-        //       Decide when the need arises.
+        var problemDetails = await _preCreateChecker.CanSaveAsync(accountId, cancellationToken);
 
-        var account = await _accountRepository.GetAccountOrDefaultAsync(id, cancellationToken);
+        if (problemDetails is not null)
+        {
+            return EnrichedResult.Fail<bool>(problemDetails);
+        }
+
+        var account = await _accountRepository.GetAccountOrDefaultAsync(accountId, cancellationToken);
 
         if (account is null)
         {
-            return false;
+            return EnrichedResult.Success(false);
         }
 
         _accountRepository.Delete(account);
 
         await _accountRepository.SaveAsync(cancellationToken);
 
-        return true;
+        return EnrichedResult.Success(true);
     }
 }
